profile changes

+++ cfn-resources/cloud-backup-snapshot/test/cfn-test-create-inputs.sh
@@ -25,49 +25,47 @@ projectName="${1}"
 clusterName=$projectName
 echo "Came inside create inputs to test"
 
-#projectId=$(atlas projects list --output json | jq --arg NAME "${projectName}" -r '.results[] | select(.name==$NAME) | .id')
-#if [ -z "$projectId" ]; then
-#    projectId=$(atlas projects create "${projectName}" --output=json | jq -r '.id')
-#    echo -e "Cant find project \"${projectName}\"\n"
-#fi
-#export MCLI_PROJECT_ID=$projectId
-#
-#clusterId=$(atlas clusters list --projectId "${projectId}"  --output json | jq --arg NAME "${clusterName}" -r '.results[]? | select(.name==$NAME) | .id')
-#if [ -z "$clusterId" ]; then
-#  echo "creating cluster.."
-#  clusterId=$(atlas clusters create "${clusterName}" --projectId "${projectId}" --backup --provider AWS --region US_EAST_1 --members 3 --tier M10 --mdbVersion 5.0 --diskSizeGB 10 --output=json | jq -r '.id')
-#fi
-#
-#status=""
-#while [[ "${status}" != "IDLE" ]]; do
-#        sleep 30
-#        status=$(atlas clusters describe "${clusterName}" --projectId "${projectId}"  --output=json | jq -r '.stateName')
-#        if [ -z "$status" ]; then
-#          status="timeout"
-#        fi
-#        echo "status: ${status}"
-#done
+projectId=$(atlas projects list --output json | jq --arg NAME "${projectName}" -r '.results[] | select(.name==$NAME) | .id')
+if [ -z "$projectId" ]; then
+    projectId=$(atlas projects create "${projectName}" --output=json | jq -r '.id')
+    echo -e "Cant find project \"${projectName}\"\n"
+fi
+export MCLI_PROJECT_ID=$projectId
 
-#echo -e "Created Cluster \"${clusterName}\" with id: ${clusterId}\n"
+clusterId=$(atlas clusters list --projectId "${projectId}"  --output json | jq --arg NAME "${clusterName}" -r '.results[]? | select(.name==$NAME) | .id')
+if [ -z "$clusterId" ]; then
+  echo "creating cluster.."
+  clusterId=$(atlas clusters create "${clusterName}" --projectId "${projectId}" --backup --provider AWS --region US_EAST_1 --members 3 --tier M10 --mdbVersion 5.0 --diskSizeGB 10 --output=json | jq -r '.id')
+fi
 
-projectId="63f35112dfbe3b111047b7fc"
-clusterName="cfn-test-bot-4ae7a4-cloud-backup-snapshot"
-#if [ -z "$clusterId" ]; then
-#    echo -e "Error Can't find Cluster \"${clusterName}\""
-#    exit 1
-#fi
+status=""
+while [[ "${status}" != "IDLE" ]]; do
+        sleep 30
+        status=$(atlas clusters describe "${clusterName}" --projectId "${projectId}"  --output=json | jq -r '.stateName')
+        if [ -z "$status" ]; then
+          status="timeout"
+        fi
+        echo "status: ${status}"
+done
+
+echo -e "Created Cluster \"${clusterName}\" with id: ${clusterId}\n"
+
+if [ -z "$clusterId" ]; then
+    echo -e "Error Can't find Cluster \"${clusterName}\""
+    exit 1
+fi
 
 rm -rf inputs
 mkdir inputs
 jq --arg group_id "$projectId" \
    --arg clusterName "$clusterName" \
-   '.ClusterName?|=$clusterName |.GroupId?|=$group_id' \
+   '.ClusterName?|=$clusterName |.ProjectId?|=$group_id' \
    "$(dirname "$0")/inputs_1_create.template.json" > "inputs/inputs_1_create.json"
 
 clusterName="${clusterName}- more B@d chars !@(!(@====*** ;;::"
 jq --arg group_id "$projectId" \
    --arg clusterName "$clusterName" \
-   '.ClusterName?|=$clusterName |.GroupId?|=$group_id' \
+   '.ClusterName?|=$clusterName |.ProjectId?|=$group_id' \
    "$(dirname "$0")/inputs_1_invalid.template.json" > "inputs/inputs_1_invalid.json"
 
 echo "mongocli iam projects delete ${projectId} --force"
